Update homekit.markdown
In the delay after start section, it might be worth adding some clarification that a delay seems to be needed for MQTT devices to load/show in HomeKit. This is needed, I assume, with all MQTT devices, but it’s definitely needed to get Ring Alarm devices showing (at least when using the Ring Alarm add-on in Hassio). And there currently isn’t any documentation that mentions these issues. 

I’m unsure if there isn’t a more elegant solution using “...network_ready,” but a 10 second delay works in my setup.